requirements.txt
blinker==1.7.0
cachetools==5.3.2
click==8.1.7
Flask==3.0.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
importlib-metadata==7.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
owlrl==6.0.2
packaging==23.2
ply==3.11
prettytable==3.9.0
PyLD==2.0.3
pyparsing==3.1.1
pyshacl==0.25.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
rdflib==7.0.0
six==1.16.0
typing_extensions==4.9.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
zipp==3.17.0
